+++ src/main/kotlin/com/github/gradle/node/exec/ExecRunner.kt
@@ -3,6 +3,7 @@ package com.github.gradle.node.exec
 import com.github.gradle.node.NodeExtension
 import com.github.gradle.node.util.ProjectApiHelper
 import org.gradle.api.file.DirectoryProperty
+import org.gradle.process.ExecOperations
 import org.gradle.process.ExecResult
 import java.io.File
 
@@ -42,8 +43,20 @@ fun computeWorkingDir(nodeProjectDir: DirectoryProperty, execConfiguration: Exec
  * different meaning to its values.
  */
 class ExecRunner {
+    @Deprecated(message = ProjectApiHelper.DEPRECATION_STRING)
     fun execute(projectHelper: ProjectApiHelper, extension: NodeExtension, execConfiguration: ExecConfiguration): ExecResult {
         return projectHelper.exec {
+            executable = execConfiguration.executable
+            args = execConfiguration.args
+            environment = computeEnvironment(execConfiguration)
+            isIgnoreExitValue = execConfiguration.ignoreExitValue
+            workingDir = computeWorkingDir(extension.nodeProjectDir, execConfiguration)
+            execConfiguration.execOverrides?.execute(this)
+        }
+    }
+
+    fun execute(execOperations: ExecOperations, extension: NodeExtension, execConfiguration: ExecConfiguration): ExecResult {
+        return execOperations.exec {
             executable = execConfiguration.executable
             args = execConfiguration.args
             environment = computeEnvironment(execConfiguration)