+++ src/main/kotlin/com/github/gradle/node/pnpm/exec/PnpmExecRunner.kt
@@ -12,30 +12,39 @@ import com.github.gradle.node.variant.VariantComputer
 import com.github.gradle.node.variant.computeNodeExec
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ProviderFactory
+import org.gradle.process.ExecOperations
 import org.gradle.process.ExecResult
 import javax.inject.Inject
 
 abstract class PnpmExecRunner {
     @get:Inject
+    abstract val execOperations: ExecOperations
+
+    @get:Inject
     abstract val providers: ProviderFactory
 
+    @Deprecated(message = ProjectApiHelper.DEPRECATION_STRING)
     fun executePnpmCommand(project: ProjectApiHelper, extension: NodeExtension, nodeExecConfiguration: NodeExecConfiguration, variants: VariantComputer): ExecResult {
+        return executePnpmCommand(extension, nodeExecConfiguration, variants)
+    }
+
+    fun executePnpmCommand(extension: NodeExtension, nodeExecConfiguration: NodeExecConfiguration, variants: VariantComputer): ExecResult {
         val npmExecConfiguration = NpmExecConfiguration("pnpm"
         ) { variantComputer, nodeExtension, pnpmBinDir -> variantComputer.computePnpmExec(nodeExtension, pnpmBinDir) }
 
-        return executeCommand(project, extension, NpmProxy.addProxyEnvironmentVariables(extension.nodeProxySettings.get(), nodeExecConfiguration),
+        return executeCommand(extension, NpmProxy.addProxyEnvironmentVariables(extension.nodeProxySettings.get(), nodeExecConfiguration),
             npmExecConfiguration,
             variants)
     }
 
-    private fun executeCommand(project: ProjectApiHelper, extension: NodeExtension, nodeExecConfiguration: NodeExecConfiguration,
+    private fun executeCommand(extension: NodeExtension, nodeExecConfiguration: NodeExecConfiguration,
                                pnpmExecConfiguration: NpmExecConfiguration,
                                variantComputer: VariantComputer): ExecResult {
         val execConfiguration =
             computeExecConfiguration(extension, pnpmExecConfiguration, nodeExecConfiguration, variantComputer).get()
         val execRunner = ExecRunner()
 
-        return execRunner.execute(project, extension, execConfiguration)
+        return execRunner.execute(execOperations, extension, execConfiguration)
     }
 
     private fun computeExecConfiguration(extension: NodeExtension, pnpmExecConfiguration: NpmExecConfiguration,